refactor: 매출 생성 컴포넌트 분리 및 토글 버튼 구현
